requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.9
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
requests>=2.31.0
google-generativeai>=0.5.2
scikit-learn>=1.4.2
pandas>=2.2.2
